pyproject.toml
[project]
name = "hepex-analysisops-agents"
version = "0.1.0"
description = "HEPEx AnalysisOps Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "google-adk>=1.22.0",
    "scipy>=1.16.3",
    "uvicorn>=0.38.0",
    "uproot>=5.0.0",
    "awkward>=2.0.0",
    "vector>=1.7.0",
    "ipykernel>=7.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

